+++ src/main/java/roomescape/auth/controller/LoginController.java
@@ -1,4 +1,4 @@
-package roomescape.member.controller;
+package roomescape.auth.controller;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
@@ -7,12 +7,12 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.auth.LoginMember;
 import roomescape.auth.dto.request.TokenRequest;
+import roomescape.auth.dto.response.LoginCheckResponse;
 import roomescape.auth.dto.response.TokenResponse;
 import roomescape.auth.service.AuthService;
+import roomescape.cofig.LoginMember;
 import roomescape.member.Member;
-import roomescape.member.dto.response.MemberResponse;
 
 @RestController
 public class LoginController {
@@ -39,8 +39,8 @@ public class LoginController {
     }
 
     @GetMapping("/login/check")
-    public ResponseEntity<MemberResponse> checkLogin(@LoginMember Member member) {
-        final MemberResponse response = new MemberResponse(member.getId(), member.getName());
+    public ResponseEntity<LoginCheckResponse> checkLogin(@LoginMember Member member) {
+        final LoginCheckResponse response = new LoginCheckResponse(member.getName());
         return ResponseEntity.ok(response);
     }
 }
