+++ src/main/java/roomescape/auth/dto/response/LoginCheckResponse.java
@@ -1,0 +1,4 @@
+package roomescape.auth.dto.response;
+
+public record LoginCheckResponse(String name) {
+}
