+++ src/main/java/roomescape/cofig/LoginMemberArgumentReslover.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.cofig;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
