+++ src/main/java/roomescape/cofig/WebConfig.java
@@ -5,8 +5,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.auth.AdminCheckInterceptor;
-import roomescape.auth.LoginMemberArgumentReslover;
 import roomescape.member.dao.MemberDao;
 import roomescape.member.infra.JwtProvider;
